# api/internal/runtime/static/openapi.yaml
openapi: 3.0.3
info:
  title: NAAP Analytics API
  version: "1"
  description: |
    Read-only analytics API for the Livepeer AI network.
    Backed by ClickHouse materialized views fed from two Kafka topics:
    `network_events` (daydream org) and `streaming_events` (cloudspe org).

    **Rate limiting:** 100 req/s sustained, burst 200 per IP.
    Exceeded requests receive `429 Too Many Requests`.

    **Default time window:** last 24 hours when `start`/`end` are omitted.

    **WEI values** (payment amounts) are returned as decimal strings to avoid
    float64 precision loss (e.g. `"1234500000000000000"`).
servers:
  - url: https://naap-api.cloudspe.com
    description: Production
  - url: http://localhost:8000
    description: Local development
  - url: https://naap-api.livepeer.cloud
    description: Production
...
components:
  parameters:
    org:
      name: org
      in: query
      description: Filter to a single organisation. Omit for all orgs.
      schema:
        type: string
        enum: [daydream, cloudspe]
    start:
      name: start
      in: query
      description: Window start (RFC3339). Defaults to 24 hours ago.
      schema:
        type: string
        format: date-time
        example: "2026-03-01T00:00:00Z"
    end:
      name: end
      in: query
      description: Window end (RFC3339). Defaults to now.
      schema:
        type: string
        format: date-time
        example: "2026-03-24T23:59:59Z"
    limit:
      name: limit
      in: query
      description: Maximum rows to return (1–1000).
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
    offset:
      name: offset
      in: query
      description: Pagination offset.
      schema:
        type: integer
        minimum: 0
        default: 0
    pipeline:
      name: pipeline
      in: query
      description: Filter by pipeline name (e.g. `text-to-image`).
      schema:
        type: string
    orch_address:
      name: orch_address
      in: query
      description: Filter by orchestrator Ethereum address (lowercased).
      schema:
        type: string
        example: "0xabc123def456"
...
    stream_id:
      name: stream_id
      in: query
      schema:
        type: string
...
  schemas:
...
    Problem:
      description: RFC 7807 Problem Details error response.
      type: object
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: "Too Many Requests"
        status:
          type: integer
          example: 429
        detail:
          type: string
          example: "rate limit exceeded"
...
    StreamStatusSample:
      type: object
      properties:
        SampleTS: { type: string, format: date-time }
        Org: { type: string }
        StreamID: { type: string }
        Gateway: { type: string }
        OrchAddress: { type: string }
        Pipeline: { type: string }
        State: { type: string }
        OutputFPS: { type: number, format: double }
        InputFPS: { type: number, format: double }
        E2ELatencyMS: { type: number, format: double }
        IsAttributed: { type: boolean }
    StreamDetail:
      type: object
      properties:
        StreamID: { type: string }
        Org: { type: string }
        Pipeline: { type: string }
        Gateway: { type: string }
        OrchAddress: { type: string }
        State: { type: string }
        StartedAt: { type: string, format: date-time }
        LastSeen: { type: string, format: date-time }
        IsClosed: { type: boolean }
        HasFailure: { type: boolean }
        Samples:
          type: array
          items: { $ref: "#/components/schemas/StreamStatusSample" }
    AttributionSummary:
      type: object
      properties:
        StartTime: { type: string, format: date-time }
        EndTime: { type: string, format: date-time }
        TotalSamples: { type: integer, format: int64 }
        AttributedCount: { type: integer, format: int64 }
        UnattributedCount: { type: integer, format: int64 }
        AttributionRate: { type: number, format: double, description: "Fraction of samples with a resolved orch address (0–1)." }

    # ── E2E Latency (R9) ────────────────────────────────────────────────────
...
  responses:
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
paths:
...
  /v1/streams/samples:
    get:
      summary: Stream status samples
      description: Returns recent per-stream telemetry samples from the last 30 days.
      operationId: listStreamSamples
      tags: [Streams]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/stream_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StreamStatusSample" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/streams/attribution:
    get:
      summary: Stream attribution summary
      description: Returns the fraction of stream samples that were successfully attributed to an orchestrator.
      operationId: getAttributionSummary
      tags: [Streams]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttributionSummary" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/streams/{stream_id}:
    get:
      summary: Stream detail
      description: Returns the full state and sample history for one stream.
      operationId: getStreamDetail
      tags: [Streams]
      parameters:
        - name: stream_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── E2E Latency (R9) ────────────────────────────────────────────────────
...